test: Extract Spring test helpers to a helpers package
The helpers package can be included in any test project without side effects like deployed routes.
