+++ flow-tests/vaadin-spring-tests/test-spring-helpers/src/test/java/com/vaadin/flow/spring/test/SpringBootOnly.java
