+++ flow-tests/vaadin-spring-tests/test-spring-security-flow-reverseproxy/src/main/resources/application.properties
@@ -1,3 +1,5 @@
 server.port=8888
 logging.level.org.springframework.security=DEBUG
 server.servlet.session.persistent=false
+proxy.port=9999
+proxy.path=/public/path
